fix: decode episodeCount as float64 to match live Podimo response
The live Podimo API returns episodeCount as a JSON number that decodes
as a float (e.g. 131.0), not an int. The int field caused a JSON decode
error that silently triggered the minimal fallback on every /subscriptions
call, losing the episode count and latest-episode date.

Change EpisodeCount to float64 and render with printf %.0f in the
template so the UI shows '131' not '131.0'.

+++ podimo/types.go
@@ -54,7 +54,7 @@ type FollowedPodcast struct {
 	ID            string        `json:"id"`
 	Title         string        `json:"title"`
 	CoverImageURL string        `json:"coverImageUrl"`
-	EpisodeCount  int           `json:"episodeCount"`
+	EpisodeCount  float64       `json:"episodeCount"`
 	LatestEpisode latestEpisode `json:"latestEpisode"`
 }
